flask_server/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.24.2 
pyyaml>=6.0   
requests>=2.23.0
opencv-python-headless>=4.8.0.76  
Flask>=2.2.3
flask-cors>=3.0.10
